MVP-6453-patch: clean up unneccessory actions / extract test constants
